Fixed issue #87 - java.lang.RuntimeException: Uncompilable source code - clone() in org.jgap.util.ICloneable is defined in an inaccessible class or interface
Netbeans actually treated it as a compilation error. Don't know what was the specific cause of this complaint from it. Maybe the JDK version it uses? If someone knows the reason why that error was raised and if it was valid then let me know. For now I'm commiting a cast to ICloneable before calling the clone() method which removes the problem for me.