Some fairly major speed improvements.
Waypoints with no requirements are no longer evaluated. This increases the overall speed by (4 - waypoints) * 100%.

Seemingly unnecessary code blocks removed in isSatisfied and getSumStuff. I tested many builds with and without waypoints and the only side-effect was increased speed.

This also makes it easier to add dynamic waypoints in the future.


note that overdroning _appears_ to be slower now, but that is only because it is no longer excluding excess waypoints (because they lack the correct drone count). It actually runs at the same speed, but not overdroning in fact runs much quicker.