Fixed some minor issues with the enforce worker parity options. Added a proper error message when drone requirement is not met.
Added dynamic overlord scoring to EcStandardFitness. Overlord score is a function of the number of overlords and the supply. This appears to stop builds that have excess overlords, now it will prefer a closer supply.

It probably needs to be tweaked, but it worked well for the builds I knew of that had the massing overlord problems, and standard builds