Fixed issue #218
The log attribute has been changed to transient.

Bug: Class com.fray.evo.EcEvolver defines non-transient non-serializable instance field log
Pattern id: SE_BAD_FIELD, type: Se, category: BAD_PRACTICE


This Serializable class defines a non-primitive instance field which is neither transient, Serializable, or java.lang.Object, and does not appear to implement the Externalizable interface or the readObject() and writeObject() methods.  Objects of this class will not be deserialized correctly if a non-Serializable object is stored in this field.